requirements.txt
discord.py[voice]
Lib